% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.R
\docType{data}
\name{mice}
\alias{mice}
\title{A simulated mice data}
\format{A data frame with 240 rows and 37 variables:
\describe{
  \item{id}{mouse ID}
  \item{time}{observation time}
  \item{group}{treatment group}
  \item{OTU.1}{relative abundance of an OTU}
  \item{OTU.2}{relative abundance of an OTU}
  \item{OTU.3}{relative abundance of an OTU}
  \item{OTU.4}{relative abundance of an OTU}
  \item{OTU.5}{relative abundance of an OTU}
  \item{OTU.6}{relative abundance of an OTU}
  \item{OTU.7}{relative abundance of an OTU}
  \item{OTU.8}{relative abundance of an OTU}
  \item{OTU.9}{relative abundance of an OTU}
  \item{OTU.10}{relative abundance of an OTU}
  \item{OTU.11}{relative abundance of an OTU}
  \item{OTU.12}{relative abundance of an OTU}
  \item{OTU.13}{relative abundance of an OTU}
  \item{OTU.14}{relative abundance of an OTU}
  \item{OTU.15}{relative abundance of an OTU}
  \item{OTU.16}{relative abundance of an OTU}
  \item{OTU.17}{relative abundance of an OTU}
  \item{OTU.18}{relative abundance of an OTU}
  \item{OTU.19}{relative abundance of an OTU}
  \item{OTU.20}{relative abundance of an OTU}
  \item{OTU.21}{relative abundance of an OTU}
  \item{OTU.22}{relative abundance of an OTU}
  \item{OTU.23}{relative abundance of an OTU}
  \item{OTU.24}{relative abundance of an OTU}
  \item{OTU.25}{relative abundance of an OTU}
  \item{OTU.26}{relative abundance of an OTU}
  \item{OTU.27}{relative abundance of an OTU}
  \item{OTU.28}{relative abundance of an OTU}
  \item{OTU.29}{relative abundance of an OTU}
  \item{OTU.30}{relative abundance of an OTU}
  \item{OTU.31}{relative abundance of an OTU}
  \item{OTU.32}{relative abundance of an OTU}
  \item{OTU.33}{relative abundance of an OTU}
  \item{OTU.34}{relative abundance of an OTU}
}}
\usage{
mice
}
\description{
A dataset containing longitudinal mice OTU table.
}
\keyword{datasets}

