% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MicroTDC.R
\name{MicroTDC}
\alias{MicroTDC}
\title{Microbial Temporal Dependency Comparison}
\usage{
MicroTDC(ana, id.var, cov.var, time.var, error_rate = 0.1,
  pct_threshold = 20, method = "spearman", dist.type = "F", heatmap = T,
  classify = F)
}
\arguments{
\item{ana}{an phyloseq object with counts data}

\item{id.var}{subject (id) variable name}

\item{cov.var}{covariates variable name}

\item{time.var}{time variable name}

\item{error_rate}{error rate percentage for relative abundance}

\item{pct_threshold}{occurance percentage threshold}

\item{method}{the correlation method ("pearson","kendall","spearman")}

\item{dist.type}{}

\item{heatmap}{whether to draw heatmap}

\item{classify}{whether to draw classifier tree}

\item{threshold}{the criteria for screening OTUs (percent of observations
large than 0.1% of minimal overall counts)}
}
\description{
Microbial Temporal Dependency Comparison
}
\examples{
#Not Run
#load("../microdata/laura.Rdata")
#ana  <- subset_samples(genus.count , Experiment == "Trans1" & Time >= 2 & Time <=34  )
#map <- sample_data(ana)
#MicroTDC(ana, id.var = "Mouse", cov.var = "Group", time.var = "Time")
}

