% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINT.R
\name{MINT}
\alias{MINT}
\title{Mcrobial Interdependence Non-parametric Test (MINT)}
\usage{
MINT(otu, id.var, cov.var, time.var, method = "spearman", dist.type = "F",
  heatmap = T, classify = F, fill.na = 0)
}
\arguments{
\item{otu}{a matrix of OTU table.}

\item{id.var}{a vector of subjects.}

\item{cov.var}{a vector of covariates.}

\item{time.var}{a vector of time variable.}

\item{method}{an option of the correlation method ("pearson","kendall","spearman"). The default method is "spearman".}

\item{dist.type}{the distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{heatmap}{a logical value indicating whether to draw heatmap. The default value is TRUE.}

\item{classify}{a logical value indicating whether to draw classifier tree. The default value is FALSE.}

\item{fill.na}{a number between 0 and 1 to fill the missing value. The default value is 0.}
}
\value{
This function returns typical, but limited, output for analysis of variance (general linear models).
\item{aov.tab}{Typical AOV table showing sources of variation, degrees of freedom, sequential sums of squares, mean squares, F statistics, partial R-squared and P values, based on N permutations.}
\item{coefficients}{matrix of coefficients of the linear model, with rows representing sources of variation and columns representing species; each column represents a fit of a species abundance to the linear model. These are what you get when you fit one species to your predictors. These are NOT available if you supply the distance matrix in the formula, rather than the site x species matrix}
\item{coef.sites}{matrix of coefficients of the linear model, with rows representing sources of variation and columns representing sites; each column represents a fit of a sites distances (from all other sites) to the linear model. These are what you get when you fit distances of one site to your predictors.}
\item{f.perms}{an N by m matrix of the null F statistics for each source of variation based on N permutations of the data. The permutations can be inspected with permustats and its support functions.}
\item{model.matrix}{The model.matrix for the right hand side of the formula.}
\item{terms}{The terms component of the model.}
}
\description{
Mcrobial Interdependence Non-parametric Test (MINT)
}
\examples{
load("data/mice.Rdata")
otu <- mice[, - (1:3)]
id.var   <- mice$id
cov.var  <- mice$group
time.var <- mice$time
MINT(otu, id.var, cov.var, time.var)

}

