% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMIT.R
\name{NMIT_phyloseq}
\alias{NMIT_phyloseq}
\title{Nonparametric Microbial Interdependence Test  (NMIT) using phyloseq data structure}
\usage{
NMIT_phyloseq(ana, id.var, cov.var, time.var, error.rate = 0.1,
  pct.threshold = 20, method = "kendall", dist.type = "F", heatmap = T,
  classify = F, fill.na = 0)
}
\arguments{
\item{ana}{a phyloseq object with counts/relative abundance data}

\item{id.var}{a vector of subjects.}

\item{cov.var}{a vector of covariates.}

\item{time.var}{a vector of time variable.}

\item{error.rate}{error rate percentage, the default is 0.1 percent.}

\item{pct.threshold}{occurance percentage threshold, the default percentage threshold is 20 percent.}

\item{method}{an option of the correlation method ("pearson","kendall","spearman"). The default method is "kendall".}

\item{dist.type}{character string, specifying the type of matrix norm to be computed. A character indicating the type of norm desired. The default is \code{"F"}. 
\itemize{
    \item{\code{"M"} or \code{"m"}}{specifies the maximum modulus of all the elements in \code{x}.}
    \item{\code{"O"}, \code{"o"} or \code{"1"}}{specifies the one norm, (maximum absolute column sum);}
    \item{\code{"I"} or \code{"i"}}{specifies the infinity norm (maximum absolute row sum);}
    \item{\code{"F"} or \code{"f"}}{specifies the Frobenius norm (the Euclidean norm of \code{x} treated as if it were a vector); and}
}}

\item{heatmap}{a logical value indicating whether to draw heatmap. The default value is TRUE.}

\item{classify}{a logical value indicating whether to draw classifier tree. The default value is FALSE.}

\item{fill.na}{a number between 0 and 1 to fill the missing value. The default value is 0.}
}
\value{
This function returns typical, but limited, output for analysis of variance (general linear models).
\item{aov.tab}{Typical AOV table showing sources of variation, degrees of freedom, sequential sums of squares, mean squares, F statistics, partial R-squared and P values, based on N permutations.}
\item{coefficients}{matrix of coefficients of the linear model, with rows representing sources of variation and columns representing species; each column represents a fit of a specie abundance to the linear model. These are what you get when you fit one species to your predictors. These are NOT available if you supply the distance matrix in the formula, rather than the site x species matrix}
\item{coef.sites}{matrix of coefficients of the linear model, with rows representing sources of variation and columns representing sites; each column represents a fit of a sites distances (from all other sites) to the linear model. These are what you get when you fit distances of one site to your predictors.}
\item{f.perms}{an N by m matrix of the null F statistics for each source of variation based on N permutations of the data. The permutations can be inspected with permustats and its support functions.}
\item{model.matrix}{The model.matrix for the right hand side of the formula.}
\item{terms}{The terms component of the model.}
}
\description{
Nonparametric Microbial Interdependence Test  (NMIT) using phyloseq data structure
}
\examples{
#Not Run
#load("../microdata/laura.Rdata")
#ana  <- subset_samples(genus.count , Experiment == "Trans1" & Time >= 2 & Time <=34  )
#map <- sample_data(ana)
#NMIT_phyloseq(ana, id.var = "Mouse", cov.var = "Group", time.var = "Time")
}

