% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MicroTDC.R
\name{OTUscreen}
\alias{OTUscreen}
\title{screening longitudinal OTUs}
\usage{
OTUscreen(ana, error_rate = 0.1, pct_threshold)
}
\arguments{
\item{ana}{an phyloseq object with counts data}

\item{error_rate}{error rate percentage for relative abundance}

\item{pct_threshold}{occurance percentage threshold}
}
\description{
screening longitudinal OTUs
}

