% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMIT.R
\name{OTUscreen}
\alias{OTUscreen}
\title{Screening OTUs}
\usage{
OTUscreen(ana, error_rate = 0.1, pct_threshold = 20)
}
\arguments{
\item{ana}{an phyloseq object with counts/relative abundance OTU table}

\item{error_rate}{error rate percentage, the default is 0.1 percent}

\item{pct_threshold}{occurance percentage threshold, the default percentage threshold is 20 percent}
}
\value{
an phyloseq object with major taxa
}
\description{
Keep major taxa with a predefined error rate and percentage threshold
}

