% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MicroTDC.R
\name{tscor}
\alias{tscor}
\title{Longitudinal correlation by samples}
\usage{
tscor(ana, method = "kendall", splitcat = "Mouse", normfun, fill.na = 0)
}
\arguments{
\item{ana}{an phyloseq object with counts data}

\item{method}{the correlation method ("pearson","kendall","spearman")}

\item{fill.na}{what value to fill if correlation is NA}

\item{split}{varialbe}

\item{nromfun}{the function to normalize raw counts}
}
\description{
Longitudinal correlation by samples
}

